requirements.txt
discord.py>=2.0.0
httpx>=0.27.2
psutil>=6.1.0
cmyui>=1.9.3
g4f==0.3.2.2
g4f[all]